test: split nexus meta-backend tests into a dedicated file
The three nexus meta-backend tests in tests/test_backends.py could not
be collected unless the optional braket extra was installed (the file
eagerly imports BraketClientBackend at module scope). Move them into
tests/test_nexus_meta_backend.py so they run in any env that has the
nexus extra.
