+++ tests/test_backends.py
@@ -17,11 +17,9 @@ from qibo_cloud_backends import (
     BraketClientBackend,
     IonQClientBackend,
     MetaBackend,
-    NexusClientBackend,
     QiboClientBackend,
     QiskitClientBackend,
 )
-import qibo_cloud_backends.nexus_client as nexus_mod
 
 NP_BACKEND = NumpyBackend()
 QISKIT_TK = os.environ.get("IBMQ_TOKEN")
@@ -89,20 +87,6 @@ def test_set_backend(backend, token):
             get_backend(), MetaBackend.load(backend, token=token).__class__
         )
         assert get_backend().name == backend
-
-
-def test_meta_backend_load_nexus_client(monkeypatch):
-    monkeypatch.setattr(nexus_mod, "_ensure_nexus_dependencies", lambda: None)
-    backend = MetaBackend.load(client="nexus-client", platform="hseries:H2-1LE")
-    assert isinstance(backend, NexusClientBackend)
-    assert backend.name == "nexus-client"
-
-
-def test_set_backend_nexus_client(monkeypatch):
-    monkeypatch.setattr(nexus_mod, "_ensure_nexus_dependencies", lambda: None)
-    set_backend("qibo-cloud-backends", client="nexus-client", platform="hseries:H2-1LE")
-    assert isinstance(get_backend(), NexusClientBackend)
-    assert get_backend().name == "nexus-client"
 
 
 def test_list_available_backends():
